requirements.txt
matplotlib
numpy
pandas
pandas_ta
xgboost
yfinance
tweepy
ta-lib
scikit-learn
wordcloud
nltk
googletrans
python-dotenv
tqdm
tradingview-ta
python-binance
flask
flask-restful
flask-jwt-extended
flask-migrate
psycopg2-binary
flask_cors
flask-sqlalchemy
gunicorn
pytest
langchain
openai
